% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/NostalgiR.R
\name{nos.legend}
\alias{nos.legend}
\title{nos.legend Generates a text-based legend for plots.}
\usage{
nos.legend(nams, legend = 1:length(nams), width)
}
\arguments{
\item{nams}{A vector that specifies what each plot symbol represents.}

\item{legend}{A character vector of the plot symbols.}

\item{width}{A numeric value that specifies the number of characters to be printed before a line-break.}
}
\description{
Generates a text-based legend that relates plot characters to functional values or outcomes.
}
\examples{
## Generates a legend that relates the plot symbols c('a','b','c') to
## the functional values c(1,2,3) with line-breaks every 14 characters.
nos.legend(c(1,2,3),c('a','b','c'),14)
}
\author{
Hien D. Nguyen
}
\seealso{
\code{\link[txtplot]{txtplot}}.
}

